+++ CHANGELOG.md
@@ -37,7 +37,7 @@ Note that this project **does not** adhere to [Semantic Versioning](http://semve
 - We improved the error message for invalid jstyles. [#6303](https://github.com/JabRef/jabref/issues/6303)
 - We changed the section name of 'Advanced' to 'Network' in the preferences and removed some obsolete options.[#6489](https://github.com/JabRef/jabref/pull/6489)
 - We improved the context menu of the column "Linked identifiers" of the main table, by truncating their texts, if they are too long. [#6499](https://github.com/JabRef/jabref/issues/6499)
-- We merged the main table tabs in the preferences dialog[#6518](https://github.com/JabRef/jabref/pull/6518)
+- We merged the main table tabs in the preferences dialog. [#6518](https://github.com/JabRef/jabref/pull/6518)
 
 ### Fixed